[BUG] prevent templates from erroring when trying to re-register via qweb
[UPD] use fetch caching, vs try catch, to load templates

I originally did not want to re-parse the XML when loading owl templates
into the qweb engine. I didn't want to depend on that code, so instead
just depended on the error they were throwing.

Switching away from that to a better solution. Caching the fetch
requests and only loading in each template a single time.